services/ophicache/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
aiofiles==24.1.0
pyyaml==6.0.2
